The import code was unnecessarily templated on the TransactionLock.
The refactoring change strips aways the TransactionLock template type
from a collection of classes. This simplifies things a little and
enables their implementations to be moved out of header files.
