+++ lib/exchange/CMakeLists.txt
@@ -82,10 +82,12 @@ set (pstore_exchange_sources
     export_names.cpp
     export_ostream.cpp
     import_compilation.cpp
+    import_debug_line_header.cpp
     import_error.cpp
     import_fixups.cpp
     import_fragment.cpp
     import_names.cpp
+    import_names_array.cpp
     import_root.cpp
     import_rule.cpp
     import_terminals.cpp
